+++ apps/opentelemetry_api_experimental/test/otel_metrics_SUITE.erl
@@ -5,10 +5,13 @@
 -include_lib("stdlib/include/assert.hrl").
 -include_lib("common_test/include/ct.hrl").
 -include_lib("opentelemetry_api/include/opentelemetry.hrl").
--include("otel_meter.hrl").
 
 all() ->
-    [noop_metrics, macros, non_overridable].
+    [{group, integer}, {group, float}].
+
+groups() ->
+    [{integer, [shuffle, parallel], [sync_instruments]},
+     {float, [shuffle, parallel], [sync_instruments]}].
 
 init_per_suite(Config) ->
     application:load(opentelemetry_api),
@@ -17,31 +20,74 @@ init_per_suite(Config) ->
 end_per_suite(_Config) ->
     ok.
 
-noop_metrics(_Config) ->
+init_per_group(Group, Config) ->
+    [{value_type, Group} | Config].
+
+end_per_group(_, _) ->
+    ok.
+
+sync_instruments(Config) ->
+    ValueType = ?config(value_type, Config),
+
     Meter = opentelemetry_experimental:get_meter(),
     ?assertMatch({otel_meter_noop, _}, Meter),
 
-    ?assert(otel_counter:new(Meter, <<"noop-measure-1">>, #{description => <<"some description">>})),
+    CounterName = a_counter,
+    CounterDesc = <<"counter description">>,
+    CounterUnit = kb,
+
+    Counter = otel_meter:create_counter(Meter, CounterName, ValueType,
+                                        #{description => CounterDesc,
+                                          unit => CounterUnit}),
+    ?assertEqual(#{meter => {otel_meter_noop,[]},
+                   module => otel_meter_noop,
+                   name => CounterName,
+                   description => CounterDesc,
+                   kind => counter,
+                   value_type => ValueType,
+                   unit => CounterUnit}, Counter),
+
+    ?assertEqual(ok, otel_counter:add(Counter, value(ValueType), #{})),
+
+    UpDownCounterName = a_updown_counter,
+    UpDownCounterDesc = <<"updown counter description">>,
+    UpDownCounterUnit = '1',
+
+    UpDownCounter = otel_meter:create_updown_counter(Meter, UpDownCounterName, ValueType,
+                                        #{description => UpDownCounterDesc,
+                                          unit => UpDownCounterUnit}),
+    ?assertEqual(#{meter => {otel_meter_noop,[]},
+                   module => otel_meter_noop,
+                   name => UpDownCounterName,
+                   description => UpDownCounterDesc,
+                   kind => updown_counter,
+                   value_type => ValueType,
+                   unit => UpDownCounterUnit}, UpDownCounter),
+
+    ?assertEqual(ok, otel_counter:add(Counter, 0 - value(ValueType), #{})),
+
+    HistogramName = a_histogram,
+    HistogramDesc = <<"histogram description">>,
+    HistogramUnit = ms,
+
+    Histogram = otel_meter:create_histogram(Meter, HistogramName, ValueType,
+                                        #{description => HistogramDesc,
+                                          unit => HistogramUnit}),
+    ?assertEqual(#{meter => {otel_meter_noop,[]},
+                   module => otel_meter_noop,
+                   name => HistogramName,
+                   description => HistogramDesc,
+                   kind => histogram,
+                   value_type => ValueType,
+                   unit => HistogramUnit}, Histogram),
+
+    ?assertEqual(ok, otel_histogram:record(Counter, value(ValueType), #{})),
 
     ok.
 
-macros(_Config) ->
-    ?otel_new_instruments([#{name => <<"macros-measure-1">>,
-                           description => <<"some description">>,
-                           kind => counter,
-                           label_keys => [],
-                           monotonic => true,
-                           absolute => true,
-                           unit => one}]),
-    ok.
+%%
 
-%% checks that opts from the user can't override static attributes of an instrument
-non_overridable(_Config) ->
-    {_, _, Instrument} = otel_counter:definition(<<"noop-measure-1">>, #{description => <<"some description">>,
-                                                                       monotonic => false,
-                                                                       synchronous => false}),
-
-    ?assert(maps:get(monotonic, Instrument)),
-    ?assert(maps:get(synchronous, Instrument)),
-
-    ok.
+value(integer) ->
+    3;
+value(float) ->
+    2.0.