+++ CHANGELOG.md
@@ -33,7 +33,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- [Allow custom text propagator to be configured via application env](https://github.com/open-telemetry/opentelemetry-erlang/pull/408)
+- [Allow custom text propagator to be configured via application
+  env](https://github.com/open-telemetry/opentelemetry-erlang/pull/408)
+- [No longer grow export table in batch processor if no export table is
+  configured](https://github.com/open-telemetry/opentelemetry-erlang/pull/413)
 
 ### [Exporter]
 
@@ -45,6 +48,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `opentelemetry:instrumentation_scope/3` or
   `opentelemetry:instrumentation_library/3` to create an `instrumentation_scope`
   record.
+
+### [Experimental API/SDK]
+
+#### Added
+
+- [Initial work to support the stable spec for the Metrics API and
+  SDK](https://github.com/open-telemetry/opentelemetry-erlang/pull/412)
 
 ## SDK 1.0.5 - 2022-05-20